Move FlashEvent from Cart to Resource component
