+++ src/Sylius/Bundle/CartBundle/Controller/CartController.php
@@ -12,8 +12,8 @@
 namespace Sylius\Bundle\CartBundle\Controller;
 
 use Sylius\Component\Cart\Event\CartEvent;
-use Sylius\Component\Cart\Event\FlashEvent;
 use Sylius\Component\Cart\SyliusCartEvents;
+use Sylius\Component\Resource\Event\FlashEvent;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
