+++ src/Sylius/Bundle/CartBundle/spec/Sylius/Bundle/CartBundle/EventListener/FlashListenerSpec.php
@@ -13,8 +13,8 @@ namespace spec\Sylius\Bundle\CartBundle\EventListener;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Sylius\Component\Cart\Event\FlashEvent;
 use Sylius\Component\Cart\SyliusCartEvents;
+use Sylius\Component\Resource\Event\FlashEvent;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBag;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Translation\TranslatorInterface;
@@ -58,9 +58,7 @@ class FlashListenerSpec extends ObjectBehavior
             ->willReturn(null)
         ;
 
-        $this
-            ->addErrorFlash($event)
-        ;
+        $this->addErrorFlash($event);
     }
 
     function it_should_add_a_custom_success_flash_message_from_event($session, FlashEvent $event, FlashBag $flashBag)
@@ -87,9 +85,7 @@ class FlashListenerSpec extends ObjectBehavior
             ->willReturn(null)
         ;
 
-        $this
-            ->addSuccessFlash($event)
-        ;
+        $this->addSuccessFlash($event);
     }
 
     function it_should_have_a_default_error_flash_message_for_event_name(
@@ -112,7 +108,6 @@ class FlashListenerSpec extends ObjectBehavior
 
         $session
             ->getBag(Argument::exact('flashes'))
-            ->shouldBeCalled()
             ->willReturn($flashBag)
         ;
 
@@ -126,9 +121,7 @@ class FlashListenerSpec extends ObjectBehavior
             ->willReturn(null)
         ;
 
-        $this
-            ->addErrorFlash($event)
-        ;
+        $this->addErrorFlash($event);
     }
 
     function it_should_have_a_default_success_flash_message_for_event_name(
@@ -156,7 +149,6 @@ class FlashListenerSpec extends ObjectBehavior
 
         $translator
             ->trans(Argument::cetera())
-            ->shouldBeCalled()
             ->willReturn($messages[SyliusCartEvents::ITEM_ADD_COMPLETED])
         ;
 
@@ -165,8 +157,6 @@ class FlashListenerSpec extends ObjectBehavior
             ->willReturn(null)
         ;
 
-        $this
-            ->addSuccessFlash($event)
-        ;
+        $this->addSuccessFlash($event);
     }
 }
